requirements.txt
pdfplumber~=0.11.6
python-telegram-bot
img2table~=1.4.1